Switch default model to liuhaotian/llava-v1.6-mistral-7b and bump transformers to 4.37.2

+++ start_network_volume.sh
@@ -2,12 +2,23 @@
 
 echo "Worker Initiated"
 
-echo "Symlinking files from Network Volume"
-ln -s /runpod-volume /workspace
+if [[ ! -L /workspace ]]; then
+    echo "Symlinking files from Network Volume"
+    ln -s /runpod-volume /workspace
+fi
 
-echo "Starting RunPod Handler"
-export PYTHONUNBUFFERED=1
-export HF_HOME="/runpod_volume"
-source /workspace/runpod-worker-llava/venv/bin/activate
-cd /workspace/runpod-worker-llava
-python3 -u rp_handler.py
+if [ -f "/workspace/venv/bin/activate" ]; then
+    echo "Starting RunPod Handler"
+    export PYTHONUNBUFFERED=1
+    export HF_HOME="/runpod_volume"
+    source /workspace/runpod-worker-llava/venv/bin/activate
+    cd /workspace/runpod-worker-llava
+    cp /rp_handler.py .
+    python3 -u rp_handler.py
+else
+    echo "ERROR: The Python Virtual Environment (/workspace/venv/bin/activate) could not be activated"
+    echo "1. Ensure that you have followed the instructions at: https://github.com/ashleykleynhans/runpod-worker-llava/blob/main/docs/building/with-network-volume.md"
+    echo "2. Ensure that you have used the Pytorch image for the installation and NOT a LLaVA image."
+    echo "3. Ensure that you have attached your Network Volume to your endpoint."
+    echo "4. Ensure that you didn't assign any other invalid regions to your endpoint."
+fi
